requirements.txt
-i https://pypi.org/simple
alembic==1.13.2; python_version >= '3.8'
aniso8601==9.0.1
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.1.3; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
decorator==5.1.1; python_version > '3.6'
executing==2.0.1; python_version >= '3.5'
faker==26.0.0; python_version >= '3.8'
flask==2.2.5; python_version >= '3.7'
flask-bcrypt==1.0.1
flask-cors==4.0.1
flask-jwt-extended==4.6.0; python_version >= '3.7' and python_version < '4'
flask-migrate==4.0.7; python_version >= '3.6'
flask-restful==0.3.10
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.0.3; python_version < '3.13' and (platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))))
gunicorn==22.0.0; python_version >= '3.7'
importlib-metadata==8.0.0; python_version < '3.10'
importlib-resources==6.4.0; python_version < '3.9'
ipdb==0.13.9; python_version >= '2.7'
ipython==8.12.3; python_version > '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
mako==1.3.5; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
matplotlib-inline==0.1.7; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
parso==0.8.4; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
psycopg2-binary==2.9.9; python_version >= '3.7'
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
setuptools==71.1.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.31; python_version >= '3.7'
sqlalchemy-serializer==1.4.12
stack-data==0.6.3
toml==0.10.2; python_version > '3.6'
traitlets==5.14.3; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
wcwidth==0.2.13
werkzeug==2.2.2; python_version >= '3.7'
zipp==3.19.2; python_version >= '3.8'
